requirements.txt
python-telegram-bot==20.*
openai==1.*
python-dotenv==1.*
